ISOmaker-Update-Late-Command

+++ Virtual/Machine/ISO-maker/Files/For-Target/late_command.sh
@@ -12,7 +12,7 @@
  chmod a+x /var/local/start.sh /var/local/finish.sh /var/local/finish-MAX-Extras.sh /var/local/grub-menu.sh /var/local/late_command.sh 
  chmod a+x /etc/rc.local 
  rm -f /var/local/ 
- rm -f /var/local/ 
+ [[ -e /var/local/ ]] && rm -f /var/local/ 
  touch /var/local/ 
  mkdir -p /usr/share/lightdm/lightdm.conf.d /etc/systemd/system/getty@tty1.service.d 
  wget -O /etc/systemd/system/getty@tty1.service.d/override.conf https://raw.githubusercontent.com/econ-ark/econ-ark-tools/master/Virtual/Machine/ISO-maker/Files/For-Target/root/etc/systemd/system/getty@tty1.service.d/override.conf
